rename: 페이지 타이틀 받아오는 유틸 lib로 분리(#89)
