+++ src/lib/index.ts
@@ -3,3 +3,5 @@ export * from './cn'
 export * from './mappers'
 export * from './table'
 export * from './portalRoot'
+export * from './datetime'
+export * from './getPageTitle'
